+++ packages/postcss-tape/src/format-asserts.ts
@@ -1,0 +1,61 @@
+export const dashesSeparator = '----------------------------------------';
+
+export function formatCSSAssertError(testCaseLabel, testCaseOptions, err, forGithubAnnotation = false) {
+	let formatted = '';
+	formatted += `\n${testCaseLabel}\n\n`;
+
+	if (testCaseOptions.message) {
+		formatted += `message :\n  ${testCaseOptions.message}\n\n`;
+	}
+
+	if (testCaseOptions.options) {
+		try {
+			formatted += `options :\n${JSON.stringify(testCaseOptions.options, null, 2)}\n\n`;
+		} catch (_) {
+			// ignore
+		}
+	}
+
+	formatted += `output changed :\n${prettyDiff(err.message)}\n`;
+
+	if (!forGithubAnnotation) {
+		formatted += '\n' + dashesSeparator;
+	}
+
+	return formatted;
+}
+
+export function formatWarningsAssertError(testCaseLabel, testCaseOptions, actual, expected, forGithubAnnotation = false) {
+	let formatted = '';
+	formatted += `\n${testCaseLabel}\n\n`;
+
+	if (testCaseOptions.message) {
+		formatted += `message :\n  ${testCaseOptions.message}\n\n`;
+	}
+
+	if (testCaseOptions.options) {
+		try {
+			formatted += `options :\n${JSON.stringify(testCaseOptions.options, null, 2)}\n\n`;
+		} catch (_) {
+			// ignore
+		}
+	}
+
+	formatted += `unexpected or missing warnings :\n+ actual ${actual}\n- expected ${expected}\n`;
+
+	if (!forGithubAnnotation) {
+		formatted += '\n' + dashesSeparator;
+	}
+
+	return formatted;
+}
+
+function prettyDiff(assertMessage: string) {
+	const newLineRegex = /[^\\](\\n)/gm;
+	const tabRegex = /(\\t)/gm;
+	return assertMessage.replace(newLineRegex, (match, p1) => { // decode new lines in CSS
+		return match.replace(p1, ' ');
+	}).replace(tabRegex, (match, p1) => { // decode tabs in CSS
+		return match.replace(p1, ' ');
+	}).replace(/\+$/gm, '').replace(/^Expected values to be strictly equal:\n/, ''); // remove trailing + outputted by `assert`
+}