+++ plugins/postcss-base-plugin/.tape.mjs
@@ -1,4 +1,7 @@
-module.exports = {
+import postcssTape from '../../packages/postcss-tape/dist/index.mjs';
+import plugin from '@csstools/postcss-base-plugin';
+
+postcssTape(plugin)({
 	basic: {
 		message: "supports basic usage",
 	},
@@ -11,4 +14,4 @@ module.exports = {
 	example: {
 		message: "minimal example",
 	},
-};
+});