+++ plugins/postcss-custom-properties/.tape.mjs
@@ -1,7 +1,10 @@
-const importWithProps = require('./test/import');
-const assert = require('assert').strict;
+import postcssTape from '../../packages/postcss-tape/dist/index.mjs';
+import plugin from 'postcss-custom-properties';
+import { strict as assert } from 'assert';
+import postcssImport from 'postcss-import';
+import fs from 'fs';
 
-module.exports = {
+postcssTape(plugin)({
 	'basic': {
 		message: 'supports basic usage'
 	},
@@ -70,23 +73,12 @@ module.exports = {
 		expect: 'basic.import.expect.css',
 		result: 'basic.import.result.css'
 	},
-	'basic:import-js': {
-		message: 'supports { importFrom: "test/import-properties{-2}?.js" } usage',
+	'basic:import-mjs': {
+		message: 'supports { importFrom: "test/import-properties{-2}?.mjs" } usage',
 		options: {
 			importFrom: [
-				'test/import-properties.js',
-				'test/import-properties-2.js'
-			]
-		},
-		expect: 'basic.import.expect.css',
-		result: 'basic.import.result.css'
-	},
-	'basic:import-cjs': {
-		message: 'supports { importFrom: "test/import-properties{-2}?.cjs" } usage',
-		options: {
-			importFrom: [
-				'test/import-properties.cjs',
-				'test/import-properties-2.cjs'
+				'test/import-properties.mjs',
+				'test/import-properties-2.mjs'
 			]
 		},
 		expect: 'basic.import.expect.css',
@@ -235,10 +227,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.scss', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.scss', 'utf8');
+				fs.rmSync('test/export-properties.scss');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.scss', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.scss', 'utf8'));
 		}
 	},
 	'basic:export-json': {
@@ -249,10 +246,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.json', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.json', 'utf8');
+				fs.rmSync('test/export-properties.json');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.json', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.json', 'utf8'));
 		}
 	},
 	'basic:export-js': {
@@ -263,10 +265,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.js', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.js', 'utf8');
+				fs.rmSync('test/export-properties.js');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.js', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.js', 'utf8'));
 		}
 	},
 	'basic:export-mjs': {
@@ -277,10 +284,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.mjs', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.mjs', 'utf8');
+				fs.rmSync('test/export-properties.mjs');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.mjs', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.mjs', 'utf8'));
 		}
 	},
 	'basic:export-css': {
@@ -291,10 +303,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.css', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.css', 'utf8');
+				fs.rmSync('test/export-properties.css');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.css', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.css', 'utf8'));
 		}
 	},
 	'basic:export-css-to': {
@@ -305,10 +322,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.css', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.css', 'utf8');
+				fs.rmSync('test/export-properties.css');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.css', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.css', 'utf8'));
 		}
 	},
 	'basic:export-css-to-type': {
@@ -319,10 +341,15 @@ module.exports = {
 		expect: 'basic.expect.css',
 		result: 'basic.result.css',
 		before() {
-			global.__exportPropertiesString = require('fs').readFileSync('test/export-properties.css', 'utf8');
+			try {
+				global.__exportPropertiesString = fs.readFileSync('test/export-properties.css', 'utf8');
+				fs.rmSync('test/export-properties.css');
+			} catch (_) {
+				// ignore
+			}
 		},
 		after() {
-			assert.strictEqual(global.__exportPropertiesString, require('fs').readFileSync('test/export-properties.css', 'utf8'));
+			assert.strictEqual(global.__exportPropertiesString, fs.readFileSync('test/export-properties.css', 'utf8'));
 		}
 	},
 	'basic:import-is-empty': {
@@ -333,10 +360,6 @@ module.exports = {
 	},
 	'import': {
 		message: 'supports "postcss-import"',
-		options: {
-			expect: 'import.expect.css',
-			result: 'import.result.css'
-		},
-		plugin: importWithProps
+		plugins: [postcssImport(), plugin()]
 	}
-};
+});