+++ plugins/postcss-custom-properties/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Custom Properties
 
+### Unreleased
+
+- Fix `.mjs` in `importFrom` when using `export default`
+
 ### 12.1.3 (January 17, 2022)
 
 - Reset plugin state after each process. It is now safe to use the plugin multiple times for different processes or when watching.